% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmatrix.r
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Correlation Matrix Function}
\usage{
correlation_matrix(x, y)
}
\description{
This function allows you to determine which player is the oldest.
}
\examples{
most_points()

}
\keyword{points}
