% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostpoints.r
\name{most_points}
\alias{most_points}
\title{Most Points Function}
\usage{
most_points(x, y)
}
\description{
This function allows you to determine which player scored the most points.
}
\examples{
most_points()
}
\keyword{points}
