% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldestplayer.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player Function}
\usage{
oldest_player(x, y)
}
\description{
This function allows you to determine which player is the oldest.
}
\examples{
most_points()

}
\keyword{points}
